+++ src/parse.ts
@@ -1,5 +1,5 @@
-import { decode } from './encoding'
-import { hasProtocol } from './utils'
+import { decode } from "./encoding";
+import { hasProtocol } from "./utils";
 export interface ParsedURL {
   protocol?: string
   host?: string
@@ -19,55 +19,55 @@ export interface ParsedHost {
   port: string
 }
 
-export function parseURL (input: string = '', defaultProto?: string): ParsedURL {
+export function parseURL (input: string = "", defaultProto?: string): ParsedURL {
   if (!hasProtocol(input, true)) {
-    return defaultProto ? parseURL(defaultProto + input) : parsePath(input)
+    return defaultProto ? parseURL(defaultProto + input) : parsePath(input);
   }
 
-  const [protocol = '', auth, hostAndPath = ''] = (input.replace(/\\/g, '/').match(/([^:/]+:)?\/\/([^/@]+@)?(.*)/) || []).splice(1)
-  const [host = '', path = ''] = (hostAndPath.match(/([^/?#]*)(.*)?/) || []).splice(1)
-  const { pathname, search, hash } = parsePath(path)
+  const [protocol = "", auth, hostAndPath = ""] = (input.replace(/\\/g, "/").match(/([^/:]+:)?\/\/([^/@]+@)?(.*)/) || []).splice(1);
+  const [host = "", path = ""] = (hostAndPath.match(/([^#/?]*)(.*)?/) || []).splice(1);
+  const { pathname, search, hash } = parsePath(path.replace(/\/(?=[A-Za-z]:)/, ""));
 
   return {
     protocol,
-    auth: auth ? auth.substr(0, auth.length - 1) : '',
+    auth: auth ? auth.slice(0, Math.max(0, auth.length - 1)) : "",
     host,
     pathname,
     search,
     hash
-  }
+  };
 }
 
-export function parsePath (input: string = ''): ParsedURL {
-  const [pathname = '', search = '', hash = ''] = (input.match(/([^#?]*)(\?[^#]*)?(#.*)?/) || []).splice(1)
+export function parsePath (input: string = ""): ParsedURL {
+  const [pathname = "", search = "", hash = ""] = (input.match(/([^#?]*)(\?[^#]*)?(#.*)?/) || []).splice(1);
 
   return {
     pathname,
     search,
     hash
-  }
+  };
 }
 
-export function parseAuth (input: string = ''): ParsedAuth {
-  const [username, password] = input.split(':')
+export function parseAuth (input: string = ""): ParsedAuth {
+  const [username, password] = input.split(":");
   return {
     username: decode(username),
     password: decode(password)
-  }
+  };
 }
 
-export function parseHost (input: string = ''): ParsedHost {
-  const [hostname, port] = (input.match(/([^/]*)(:0-9+)?/) || []).splice(1)
+export function parseHost (input: string = ""): ParsedHost {
+  const [hostname, port] = (input.match(/([^/]*)(:0-9+)?/) || []).splice(1);
   return {
     hostname: decode(hostname),
     port
-  }
+  };
 }
 
 export function stringifyParsedURL (parsed: ParsedURL) {
-  const fullpath = parsed.pathname + (parsed.search ? (parsed.search.startsWith('?') ? '' : '?') + parsed.search : '') + parsed.hash
+  const fullpath = parsed.pathname + (parsed.search ? (parsed.search.startsWith("?") ? "" : "?") + parsed.search : "") + parsed.hash;
   if (!parsed.protocol) {
-    return fullpath
+    return fullpath;
   }
-  return parsed.protocol + '//' + (parsed.auth ? parsed.auth + '@' : '') + parsed.host + fullpath
+  return parsed.protocol + "//" + (parsed.auth ? parsed.auth + "@" : "") + parsed.host + fullpath;
 }